refactor(desktop): split hosts/windows.py god-module into a package
windows.py mixed five unrelated concerns (tray icon rendering, PCM tone
synthesis, venv-bypass autostart, single-instance mutex, resident-host
wiring) in one 687-line file, making "what happens when I press the
hotkey" require reading through unrelated autostart/VBScript code.

Split into hosts/windows/{feedback,autostart,tray,resident}.py, with
windows/__init__.py left as a thin composition point. No behavior
change; tests/test_hosts.py updated to monkeypatch the new submodule
locations instead of the flat windows module.

Closes #55
