+++ like_spotify/hosts/windows/__main__.py
@@ -1,0 +1,6 @@
+import sys
+
+from . import main
+
+if __name__ == "__main__":
+    sys.exit(main())
